docs/mmos/board/requirements.txt
click>=8.0.0
PyYAML>=6.0.0
